backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.0
pydantic-settings==2.7.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
python-multipart==0.0.18
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
httpx==0.28.0
anthropic>=0.79.0
